마커에서 자세히보기 클릭 시 장소 새부내용 페이지 이동 기능 구현

+++ CompanionAnimalProject/src/main/java/com/example/demo/model/Place.java
@@ -1,17 +1,17 @@
 package com.example.demo.model;
 
 public class Place {
-	private int service_idx;
+	private long service_idx;
 	private String category;
 	private String name;
 	private String address;
 	private double latitude;
 	private double longitude;
 	
-	public int getService_idx() {
+	public long getService_idx() {
 		return service_idx;
 	}
-	public void setService_idx(int service_idx) {
+	public void setService_idx(long service_idx) {
 		this.service_idx = service_idx;
 	}
 	public String getCategory() {
